requirements_cpu.txt
absl-py==0.12.0
alembic==1.4.1
apache-beam==2.31.0
astunparse==1.6.3
attrs==21.2.0
avro-python3==1.9.2.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
cloudpickle==1.6.0
contextlib2==21.6.0
crcmod==1.7
cycler==0.10.0
Cython==0.29.24
databricks-cli==0.14.3
dataclasses==0.8
dill==0.3.1.1
dm-tree==0.1.6
docker==5.0.0
docopt==0.6.2
entrypoints==0.3
fastavro==1.4.4
Flask==2.0.1
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gin-config==0.4.0
gitdb==4.0.7
GitPython==3.1.18
google-api-core==1.31.0
google-api-python-client==2.14.0
google-auth==1.33.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-bigquery==1.21.0
google-cloud-core==1.7.1
google-pasta==0.2.0
google-resumable-media==0.4.1
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.34.1
gunicorn==20.1.0
hdfs==2.6.0
httplib2==0.19.1
idna==3.2
importlib-metadata==4.6.1
importlib-resources==5.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kaggle==1.5.12
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lvis==0.5.3
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mlflow==1.19.0
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.1
object-detection @ file:///datadrive/experiments/tensorflow/models/research
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandas==1.1.5
Pillow==8.3.1
portalocker==2.0.0
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.2
promise==2.3
protobuf==3.17.3
psutil==5.8.0
py-cpuinfo==8.0.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pydot==1.4.2
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-editor==1.0.4
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacrebleu==1.5.1
scikit-learn==0.24.2
scipy==1.4.1
sentencepiece==0.1.96
seqeval==1.2.2
six==1.15.0
smmap==4.0.0
SQLAlchemy==1.4.22
sqlparse==0.4.1
tabulate==0.8.9
tensorflow-cpu==2.3.0
text-unidecode==1.3
threadpoolctl==2.2.0
tqdm==4.61.2
typeguard==2.12.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.6
websocket-client==1.1.0
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
